Move WorkItemProgress class into TaskProgressBase
